+++ CHANGES.rst
@@ -12,6 +12,9 @@ New features:
 * Change the way tagged fields are managed in the protocol definitions
   (pr #1162 by @vmaurin)
 
+Bugfixes:
+
+* Don't send more than one JoinGroup request when more than one Assignors are specified (pr #1168 by @ilolis)
 
 0.14.0 (2026-04-29)
 ===================